backend_fastapi/requirements.txt
# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0  # MySQL connector
# alembic>=1.12.0  # Optional for simple setup

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.1.2
python-dotenv>=1.0.0

# AI and ML
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
numpy>=1.24.0
pandas>=2.0.0

# Document processing
PyMuPDF>=1.23.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3

# Validation
pydantic>=2.5.0
pydantic-settings>=2.0.0
email-validator>=2.1.0

# CORS (FastAPI handles CORS natively)

# Utilities
python-magic>=0.4.27
chardet>=5.2.0
tqdm>=4.66.0
